# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BabyAI API
  version: 0.1.0
  description: >
    BabyAI MVP API. 육아 기록, AI 조회, 리포트, 구독/사진 공유, Siri/Bixby 연동을 포함한다.
servers:
  - url: http://localhost:8000
    description: local
...
paths:
...
  /api/v1/data/export.csv:
    get:
      tags: [settings]
      summary: Export baby event data as CSV
      operationId: exportBabyDataCsv
      parameters:
        - in: query
          name: baby_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: CSV data export
          content:
            text/csv:
              schema:
                type: string
                format: binary
...